+++ doc/changelog.rst
@@ -5,6 +5,8 @@ Changelog
 
 - Add :py:attr:`spead2.recv.Stream.allow_unsized_heaps` to support rejecting
   packets without a heap length.
+- Add extended custom memcpy support (C++ only) for scattering data from
+  packets.
 
 .. rubric:: 1.10.1