+++ doc/changelog.rst
@@ -1,7 +1,7 @@
 Changelog
 =========
 
-.. rubric:: Development version
+.. rubric:: 1.11.0
 
 - Add :py:attr:`spead2.recv.Stream.allow_unsized_heaps` to support rejecting
   packets without a heap length.